requirements.txt
requests==2.31.0
PyMySQL==1.1.0
DBUtils==3.0.3
python-dotenv==1.0.1